feat: add report generation as cli
The previous way to get reports required users to upload
their data to ObservableHQ. Based on our experience we expect
customers to be wary of sharing anything about their code with
third parties. This change allows users to create the reports
locally, without any of their data going towards any third party.

The change was kept to a minimum on purpose. There is no
archiving or serving of files from within the tool as better
tools exist for that purpose and users can use their own preference
to achieve this.